improvement: rename dto to schema and enhancement zod validation
