--- libs/entities/src/dto/auth-dto/index.ts
@@ -1,10 +1,0 @@
-export * from './forgot-password-dto';
-export * from './login-dto';
-export * from './logout-dto';
-export * from './register-dto';
-export * from './req-user.dto';
-export * from './reset-password-dto';
-export * from './forgot-password-dto';
-export * from './verify-dto';
-export * from './logout-dto';
-export * from './resend-otp-dto';
