--- libs/entities/src/dto/index.ts
@@ -1,3 +1,0 @@
-export * from './auth-dto';
-export * from './student-dto';
-export * from './user-dto';
