--- libs/entities/src/dto/student-dto/index.ts
@@ -1,2 +1,0 @@
-export * from './create-student.dto';
-export * from './update-student.dto';
