--- libs/entities/src/dto/user-dto/index.ts
@@ -1,2 +1,0 @@
-export * from './create-user.dto';
-export * from './update-user.dto';
