+++ libs/entities/src/schemas/auth-schemas/logout-schema.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
 import { IsString } from 'class-validator';
 
-export class UserDto {
+export class UserSchema {
   @ApiProperty({
     example: '2fc66b45-e6f6-48c7-89f1-b62d0a398cfc',
   })
@@ -10,12 +10,12 @@ export class UserDto {
   public sub!: string;
 }
 
-export class LogoutDto {
+export class LogoutSchema {
   @ApiProperty()
   public refresh_token!: string;
 }
 
-export const UserDtoZodSchema = z.object({
+export const UserZodSchema = z.object({
   sub: z.string(),
 });
 
@@ -23,5 +23,5 @@ export const LogoutZodSchema = z.object({
   refresh_token: z.string(),
 });
 
-export type TUserDtoSchema = z.infer<typeof UserDtoZodSchema>;
+export type TUserDtoSchema = z.infer<typeof UserZodSchema>;
 export type TLogoutSchema = z.infer<typeof LogoutZodSchema>;
