+++ libs/entities/src/schemas/index.ts
@@ -1,0 +1,3 @@
+export * from './auth-schemas';
+export * from './student-schemas';
+export * from './user-schemas';
