+++ libs/entities/src/schemas/user-schemas/create-user-schema.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
 
-export class CreateUserDto {
+export class CreateUserSchema {
   @ApiProperty()
   public email!: string;
 
@@ -37,7 +37,10 @@ export const CreateUserZodSchema = z.object({
     .nonempty({
       message: 'Password tidak boleh kosong',
     })
-    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).*$/),
+    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).*$/, {
+      message:
+        'Password harus memiliki setidaknya 6 karakter dan mengandung setidaknya 1 huruf kecil, 1 huruf besar, dan 1 angka. Tidak boleh mengandung simbol ',
+    }),
 });
 
 export type TCreateUserSchema = z.infer<typeof CreateUserZodSchema>;
