+++ libs/entities/src/schemas/user-schemas/update-user-schema.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
 
-export class UpdateUserDto {
+export class UpdateUserSchema {
   @ApiProperty()
   public email!: string;
 
@@ -43,7 +43,12 @@ export const UpdateUserZodSchema = z.object({
     .nonempty({
       message: 'Password tidak boleh kosong',
     })
-    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).*$/),
+    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).*$/, {
+      message:
+        'Password harus memiliki setidaknya 6 karakter dan mengandung setidaknya 1 huruf kecil, 1 huruf besar, dan 1 angka. Tidak boleh mengandung simbol ',
+    }),
   role_id: z.number().optional(),
   photo: z.string().optional(),
 });
+
+export type TUpdateUserSchema = z.infer<typeof UpdateUserZodSchema>;
